Added new plots

+++ vasp_parser.py
@@ -62,7 +62,7 @@ def read_infiles(directory):
     forces = np.loadtxt(dir+'forces.infile')
     positions = np.loadtxt(dir+'positions.infile')
     potentials = np.loadtxt(dir+'potentials.infile')
-    num_atoms = np.loadtxt(dir+'numatoms.infile')
+    num_atoms = np.loadtxt(dir+'numatoms.infile', dtype=int)
     
     return forces, positions, np.array(potentials), num_atoms
